Fix: make DualPipeNetConn cross-platform
Uses conditional compilation to restrict usage of syscall.Socketpair
to unix systems. The original use of a net.Pipe is used on windows.
