+++ lib/utils/pipenetconn.go
@@ -19,12 +19,8 @@ package utils
 import (
 	"io"
 	"net"
-	"os"
 	"sync"
-	"syscall"
 	"time"
-
-	"github.com/gravitational/trace"
 )
 
 // PipeNetConn implements net.Conn from a provided io.Reader,io.Writer and
@@ -101,38 +97,4 @@ func (nc *PipeNetConn) SetReadDeadline(t time.Time) error {
 
 func (nc *PipeNetConn) SetWriteDeadline(t time.Time) error {
 	return nil
-}
-
-// DualPipeNetConn creates a pipe to connect a client and a server. The
-// two net.Conn instances are wrapped in an PipeNetConn which holds the source and
-// destination addresses.
-//
-// The pipe is constructed from a syscall.Socketpair instead of a net.Pipe because
-// the synchronous nature of net.Pipe causes it to deadlock when attempting to perform
-// TLS or SSH handshakes.
-func DualPipeNetConn(srcAddr net.Addr, dstAddr net.Addr) (net.Conn, net.Conn, error) {
-	// File descriptor ownership is transfer to the net.FileConn created below.
-	fds, err := syscall.Socketpair(syscall.AF_LOCAL, syscall.SOCK_STREAM, 0)
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-
-	f1 := os.NewFile(uintptr(fds[0]), srcAddr.String())
-	client, err := net.FileConn(f1)
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-
-	f2 := os.NewFile(uintptr(fds[1]), dstAddr.String())
-	server, err := net.FileConn(f2)
-	if err != nil {
-		return nil, nil, trace.NewAggregate(err, client.Close())
-	}
-
-	// By passing the net.FileConn as closers we are ensuring that when
-	// the connection is closed the file descriptors will be cleaned up.
-	serverConn := NewPipeNetConn(server, server, server, dstAddr, srcAddr)
-	clientConn := NewPipeNetConn(client, client, client, srcAddr, dstAddr)
-
-	return serverConn, clientConn, nil
 }
