+++ lib/utils/pipenetconn_windows.go
@@ -1,0 +1,27 @@
+// Copyright 2022 Gravitational, Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package utils
+
+// DualPipeAddrConn creates a net.Pipe to connect a client and a server. The
+// two net.Conn instances are wrapped in an addrConn which holds the source and
+// destination addresses.
+func DualPipeNetConn(srcAddr net.Addr, dstAddr net.Addr) (*PipeNetConn, *PipeNetConn) {
+	server, client := net.Pipe()
+
+	serverConn := NewPipeNetConn(server, server, server, dstAddr, srcAddr)
+	clientConn := NewPipeNetConn(client, client, client, srcAddr, dstAddr)
+
+	return serverConn, clientConn
+}
